move psnr & ssim (#317)
* move psnr & ssim

* tests

* deprecate
