+++ setup.cfg
@@ -90,6 +90,12 @@ ignore_errors = True
 ignore_errors = True
 
 # todo: add proper typing to this module...
+[mypy-torchmetrics.image.ssim]
+ignore_errors = True
+[mypy-torchmetrics.image.psnr]
+ignore_errors = True
+
+# todo: add proper typing to this module...
 [mypy-torchmetrics.classification.*]
 ignore_errors = True
