+++ torchmetrics/image/__init__.py
@@ -14,3 +14,5 @@
 from torchmetrics.image.fid import FID  # noqa: F401
 from torchmetrics.image.inception import IS  # noqa: F401
 from torchmetrics.image.kid import KID  # noqa: F401
+from torchmetrics.image.psnr import PSNR  # noqa: F401
+from torchmetrics.image.ssim import SSIM  # noqa: F401
